+++ lib/batch_api/routing_helper.rb
@@ -2,11 +2,13 @@ module BatchApi
   module RoutingHelper
     DEFAULT_VERB = :post
     DEFAULT_ENDPOINT = "/batch"
+    DEFAULT_TARGET = "batch_api/batch#batch"
 
     def batch_api(options = {})
       endpoint = options.delete(:endpoint) || DEFAULT_ENDPOINT
       verb = options.delete(:via) || DEFAULT_VERB
-      match({endpoint => "batch_api/batch#batch", via: verb}.merge(options))
+      target = options.delete(:target) || DEFAULT_TARGET
+      match({endpoint => target, via: verb}.merge(options))
     end
   end
 end