+++ spec/lib/sequential_spec.rb
@@ -1,0 +1,5 @@
+require 'spec_helper'
+
+describe BatchApi::Processor::Strategies::Sequential do
+  pending "needs a test"
+end