+++ changelog.md
@@ -1,5 +1,12 @@
+v0.0.3
+* Encapsulate processing into a Processor module
+* Prepare for parallel processing in the future
+* Add specific errors
+* Allow controlling the routing target
+
 v0.0.2
 * Add config module
+* Add options for operation limit, endpoint, and verb
 
 v0.0.1
 * Initial build